requirements.txt
flask==2.2.3
fasttext-wheel==0.9.2
nltk==3.8.1
emoji==2.8.0
gunicorn==21.2.0
sentence_transformers==2.2.2